Refactor python lambda construct
Import path fixes after refactor

Deleted blank deployment readme
