+++ lib/workload/stateless/constructs/python-lambda-layer/index.ts
@@ -2,23 +2,26 @@ import { Construct } from 'constructs';
 import { PythonLayerVersion } from '@aws-cdk/aws-lambda-python-alpha';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 
-export interface LambdaLayerConstructProps {
+export interface PythonLambdaLayerConstructProps {
+  layer_name: string;
   layer_directory: string;
+  layer_description: string;
 }
 
-export class LambdaLayerConstruct extends Construct {
+export class PythonLambdaLayerConstruct extends Construct {
   public readonly lambda_layer_arn: string;
   public readonly lambda_layer_version_obj: PythonLayerVersion;
 
-  constructor(scope: Construct, id: string, props: LambdaLayerConstructProps) {
+  constructor(scope: Construct, id: string, props: PythonLambdaLayerConstructProps) {
     super(scope, id);
 
-    this.lambda_layer_version_obj = new PythonLayerVersion(this, 'bs_runs_upload_layer', {
+    this.lambda_layer_version_obj = new PythonLayerVersion(this, 'python_lambda_layer', {
+      layerVersionName: props.layer_name,
       entry: props.layer_directory,
       compatibleRuntimes: [lambda.Runtime.PYTHON_3_11],
       compatibleArchitectures: [lambda.Architecture.X86_64],
       license: 'GPL3',
-      description: 'A layer to enable the bssh manager tools layer',
+      description: props.layer_description,
       bundling: {
         commandHooks: {
           beforeBundling(inputDir: string, outputDir: string): string[] {
