+++ lib/workload/stateless/statelessStackCollectionClass.ts
@@ -14,8 +14,8 @@ import {
   SequenceRunManagerStack,
   SequenceRunManagerStackProps,
 } from './stacks/sequence-run-manager/deploy/stack';
-import { BsRunsUploadManagerStackProps } from './stacks/bs-runs-upload-manager/deploy/stacks/bs_runs_upload_manager_stack';
-import { BsRunsUploadManagerStack } from './stacks/bs-runs-upload-manager/deploy/stacks/bs_runs_upload_manager_stack';
+import { BsRunsUploadManagerStackProps } from './stacks/bs-runs-upload-manager/deploy/stack';
+import { BsRunsUploadManagerStack } from './stacks/bs-runs-upload-manager/deploy/stack';
 
 export interface StatelessStackCollectionProps {
   postgresManagerStackProps: PostgresManagerStackProps;
